Improve test coverage
