+++ packages/datasource-value/plugin.js
@@ -1,8 +1,13 @@
 'use strict'
 
 const Joi = require('joi')
+const ValueTransport = require('./src/value-transport')
 
 class ValueDatasource {
+
+  register (dashboard, options) {
+    dashboard.contributeDatasource('value', ValueTransport)
+  }
 
   get widgetValidation () {
     return Joi.object({
